Refactored code in asw/apps/xmlrpc/XMLRPCServer.h and created new classes in rmw (ROS middleware).
