+++ rmw/TopicReader.h
@@ -1,0 +1,38 @@
+#ifndef RMW_TOPICREADER_H_
+#define RMW_TOPICREADER_H_
+#define MAX_TOPIC_LEN 48
+
+#include "FreeRTOS.h"
+#include "task.h"
+#include "semphr.h"
+
+class TopicReader
+{
+	char topic[MAX_TOPIC_LEN];
+	char callerID[MAX_TOPIC_LEN];
+	uint32_t connectionID;
+	xQueueHandle qHandle;
+
+	static void onResponse(const void* obj,const char* data);
+
+	void requestTopic(const char* ip, uint16_t serverPort);
+	static void connectPublishers(const void* obj, const char* data);
+	static const int RX_QUEUE_MSG_SIZE = 128;
+	static const int MAX_CALLBACKS = 5;
+
+	void(*callbacks[MAX_CALLBACKS])(void* data, void* obj);
+	void* objects[MAX_CALLBACKS];
+
+
+public:
+	TopicReader(const char* callerID, const char* topic, const char* msgType);
+	void addCallback(void(*callback)(void* data, void* obj), void* obj);
+	static void task(void* arg);
+	const char* getTopic();
+	uint32_t getConnectionID();
+	void enqueueMessage(const char* msg);
+	void dequeueMessage(char* msg);
+};
+
+
+#endif /* RMW_TOPICREADER_H_ */
